🎨 Découpage de la page historique en helpers testables + spec recours avec date d'accord
