+++ packages/specifications/src/projet/éliminé/recours/stepDefinitions/recours.given.ts
@@ -66,7 +66,7 @@ async function créerDemandeRecours(this: PotentielWorld, dateDemande?: DateTime
   const { raison, demandéLe, demandéPar, pièceJustificative } =
     this.éliminéWorld.recoursWorld.demanderRecoursFixture.créer({
       demandéPar: this.utilisateurWorld.porteurFixture.email,
-      demandéLe: dateDemande,
+      ...(dateDemande && { demandéLe: dateDemande }),
     });
 
   await mediator.send<Éliminé.Recours.DemanderRecoursUseCase>({
